feat: ext module
