+++ h3/src/ext.rs
@@ -1,0 +1,33 @@
+//! Extensions for the HTTP/3 protocol.
+
+use std::str::FromStr;
+
+/// Describes the `:protocol` pseudo-header for extended connect
+///
+/// See: [https://www.rfc-editor.org/rfc/rfc8441#section-4]
+#[derive(Copy, PartialEq, Debug, Clone)]
+pub struct Protocol(ProtocolInner);
+
+impl Protocol {
+    /// WebTransport protocol
+    pub const WEB_TRANSPORT: Protocol = Protocol(ProtocolInner::WebTransport);
+}
+
+#[derive(Copy, PartialEq, Debug, Clone)]
+enum ProtocolInner {
+    WebTransport,
+}
+
+/// Error when parsing the protocol
+pub struct InvalidProtocol;
+
+impl FromStr for Protocol {
+    type Err = InvalidProtocol;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "webtransport" => Ok(Self(ProtocolInner::WebTransport)),
+            _ => Err(InvalidProtocol),
+        }
+    }
+}
