+++ h3/src/lib.rs
@@ -4,12 +4,12 @@
 
 pub mod client;
 pub mod error;
+pub mod ext;
 pub mod quic;
 pub(crate) mod request;
 pub mod server;
 
 pub use error::Error;
-pub use proto::headers::Protocol;
 
 mod buf;
